term-project/requirements.txt
autopep8==2.0.2
psutil==5.9.4
psycopg2-binary==2.9.6
pycodestyle==2.10.0
tomli==2.0.1
